+++ SilverBotDS/SilverBotDS.csproj
@@ -50,13 +50,13 @@
 
     <ItemGroup>
         <PackageReference Include="Analytics" Version="3.8.0" />
-        <PackageReference Include="DSharpPlus" Version="4.2.0-nightly-01062">
+        <PackageReference Include="DSharpPlus" Version="4.2.0-nightly-01064">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
-        <PackageReference Include="DSharpPlus.CommandsNext" Version="4.2.0-nightly-01062" />
-        <PackageReference Include="DSharpPlus.Interactivity" Version="4.2.0-nightly-01062" />
-        <PackageReference Include="DSharpPlus.SlashCommands" Version="4.2.0-nightly-01062" />
-        <PackageReference Include="DSharpPlus.VoiceNext" Version="4.2.0-nightly-01062">
+        <PackageReference Include="DSharpPlus.CommandsNext" Version="4.2.0-nightly-01064" />
+        <PackageReference Include="DSharpPlus.Interactivity" Version="4.2.0-nightly-01064" />
+        <PackageReference Include="DSharpPlus.SlashCommands" Version="4.2.0-nightly-01064" />
+        <PackageReference Include="DSharpPlus.VoiceNext" Version="4.2.0-nightly-01064">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
         <PackageReference Include="Fortnite-API-Wrapper" Version="2.3.2" />
@@ -74,7 +74,7 @@
         <PackageReference Include="MathParser.org-mXparser" Version="4.4.2" />
         <PackageReference Include="Microsoft.AspNetCore.Components.ProtectedBrowserStorage" Version="5.0.0-rc.1.20451.17" />
         <PackageReference Include="Microsoft.Azure.SignalR" Version="1.15.0" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.1.0-1.final" />
+        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.1.0-3.final" />
         <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.1">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
@@ -97,13 +97,13 @@
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="6.0.3" />
-        <PackageReference Include="NSwag.ApiDescription.Client" Version="13.15.5">
+        <PackageReference Include="NSwag.ApiDescription.Client" Version="13.15.7">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Radzen.Blazor" Version="3.14.9" />
+        <PackageReference Include="Radzen.Blazor" Version="3.14.12" />
         <PackageReference Include="ScottPlot" Version="4.1.32" />
-        <PackageReference Include="Serilog" Version="2.11.0-dev-01377">
+        <PackageReference Include="Serilog" Version="2.11.0-dev-01380">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
         <PackageReference Include="Serilog.AspNetCore" Version="4.1.1-dev-00250" />
@@ -124,14 +124,14 @@
         <PackageReference Include="System.Globalization.Calendars" Version="4.3.0">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
-        <PackageReference Include="System.Linq.Async" Version="5.1.0" />
+        <PackageReference Include="System.Linq.Async" Version="6.0.1" />
         <PackageReference Include="System.Net.Http" Version="4.3.4">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
         <PackageReference Include="System.Runtime" Version="4.3.1">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
-        <PackageReference Include="System.Text.Json" Version="6.0.1">
+        <PackageReference Include="System.Text.Json" Version="6.0.2-mauipre.1.22054.8">
           <TreatAsUsed>true</TreatAsUsed>
         </PackageReference>
         <PackageReference Include="System.Text.RegularExpressions" Version="4.3.1">
